refactor(protocol): split the intent graph into top-level nodes
